Create CustomMouseClick.html
This is the meat and potatos of how to handle DOM mouse clicks in Three.js  

you make a ray caster (an object that acts like a laser light pointer) that shoots out from where ever the mouse is when the user clicks.
The ray caster then returns an array of any objects that got hit by the ray caster which I stored in a variable called intersects. I only care about the first one hit so I stored it in a variable called intersect. Then I am going to store the object info in a var called newSelected. newSelected is a mesh object which you can now identify by name or uuid. In this case I select a specific object out by name and rotate the object if the user clicks it.